Refactor to eliminate offscreen permission
Keep the limit of requesting a single profile at a time, by moving the fetch of the github profile page to the content script, but using the background messaging to maintain a pseudo-lock to ensure that only a fetch of a given profile occurs at one time across all instances of the content script.

Also fixed issue where @username anchor references could display incorrect data.